requirements.txt
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
DateTime==4.3
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.3
django-crispy-forms==1.10.0
django-import-export==2.4.0
et-xmlfile==1.0.1
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
pylint==2.6.0
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.2
typed-ast==1.4.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.2.0
